requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
timm==1.0.3
numpy==1.26.4
opencv-python-headless==4.9.0.80
imagehash==4.3.1
hashlib-compat==1.0.0
pydantic==2.7.1
aiofiles==23.2.1
python-dotenv==1.0.1
rich==13.7.1
scikit-learn>=1.4.0
joblib>=1.3.0
